+++ linking-platform/link-picker/src/jest.d.ts
@@ -1,0 +1,8 @@
+export {};
+declare global {
+  namespace jest {
+    interface Matchers<R> {
+      toBeFiredWithAnalyticEventOnce(event: any, channel?: string): R;
+    }
+  }
+}