+++ media/smart-card/src/view/EmbedModal/components/link-info/link-info-button/index.tsx
@@ -24,6 +24,7 @@ const LinkInfoButton: React.FC<LinkInfoButtonProps> = ({
         href={href}
         iconBefore={icon}
         onClick={onClick}
+        spacing="none"
         target={target}
         testId={`${testId}-button`}
       />